requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.2
ldap3==2.9.1
Markdown==3.3.4
MarkupSafe==2.0.1
mysql-client==0.0.1
openapi-codec==1.3.2
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyodbc==4.0.32
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
simplejson==3.17.5
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.7
wfastcgi==3.0.0
